back-end/requirements.txt
# --- Backend y Servidor ---
fastapi
uvicorn
gunicorn
python-dotenv
mysql-connector-python

# --- Machine Learning y Ciencia de Datos ---
pandas
numpy
scipy
scikit-learn
joblib